requirements.txt
flask~=3.0.0
pymongo~=4.6.0
requests~=2.31.0
celery~=5.3.4
beautifulsoup4~=4.12.2
python-decouple~=3.8
eventlet~=0.33.3
flask-cors~=4.0.0